Add helper functions
Need to use device detection to separate out view between mobile and
desktop.  As it is now, it is very slow on mobile. All of it is due to
the autoform taking time to populate.  Therefore my plan is to provide
links on mobile (separate template) and use a differnet route for them.